refactor(NpcData+WorldData): Renamed these objects to *Container to better reflect the fact, that they _contain_ other data and object.
