+++ Assets/UnZENity-Core/Scripts/Creator/Meshes/V2/Builder/BarrierMeshBuilder.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using GUZ.Core.Data.Container;
 using GUZ.Core.Extensions;
 using GUZ.Core.Globals;
 using GUZ.Core.World;
