+++ Assets/UnZENity-Core/Scripts/Creator/Meshes/V2/Builder/TextureArrayBuilder.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using GUZ.Core.Caches;
+using GUZ.Core.Data.Container;
 using GUZ.Core.Globals;
-using GUZ.Core.World;
 using UnityEngine;
 using UnityEngine.Rendering;
 using ZenKit;
