+++ Assets/UnZENity-Core/Scripts/Data/Container/NpcContainer.cs
@@ -2,7 +2,7 @@ using GUZ.Core.Properties;
 using UnityEngine;
 using ZenKit.Daedalus;
 
-namespace GUZ.Core.Data
+namespace GUZ.Core.Data.Container
 {
     /// <summary>
     /// Container class for NPC data. It includes references to ZenKit
@@ -11,7 +11,7 @@ namespace GUZ.Core.Data
     /// and UnZENity:
     ///   1. NpcProperties - Component storing data next to its GameObject
     /// </summary>
-    public class NpcData
+    public class NpcContainer
     {
         public NpcInstance Instance;
         public ZenKit.Vobs.Npc Vob;
